% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrem.R
\name{lre_auto_bk}
\alias{lre_auto_bk}
\title{Solving an LRE model according to Blanchard and Kahn}
\usage{
lre_auto_bk(A, n)
}
\arguments{
\item{A}{Coefficient matrix for the step function}

\item{n}{Number of predetermined variables}
}
\value{
Two functions, decision rule g() and a motion function h()
}
\description{
Solving an LRE model according to Blanchard and Kahn
}
