% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrem.R
\name{lre_auto_klein}
\alias{lre_auto_klein}
\title{Solving an LRE model according to the Klein method with use of QZ decomposition}
\usage{
lre_auto_klein(A, E, n)
}
\arguments{
\item{A}{Coefficient matrix on the previous step of the difference equation}

\item{E}{Coefficient matrix of the current(next) step}

\item{n}{Number of predeterminate parameters}
}
\value{
Two functions g() and h() for a decision rule and law of motion respectively
}
\description{
Solving an LRE model according to the Klein method with use of QZ decomposition
}
